Add files via upload
This repository holds the official CodeAlpha logo, which has been created in a modern technological and minimalist design. CodeAlpha is an Ed-Tech firm that offers online training and internship courses to students who want to pursue software development careers.
I am currently doing an internship at CodeAlpha as a Graphic Design Intern, and this logo is included in my internship duties.